Updated irc.py, more specific the admin function
I fixed it so that the split operation only occurs once and the command and
first argument are immediately pulled out of the split and stored. This should
reduce the number of potential calls to split in case the interpreter doesn't
catch the fact that the calls will always be the same.

I fixed the !verbose action to just not the value already stored within
self.verbose instead of a goofy if/else statement.

I also added a check at the beginning of the admin function to check to see if
the command starts with a !. If it doesn't then end the admin check right
there. This may be redundant.